rpc: move server-related code to a separate package
