+++ pkg/rpc/server/prometheus.go
@@ -1,4 +1,4 @@
-package rpc
+package server
 
 import "github.com/prometheus/client_golang/prometheus"
