+++ pkg/rpc/server/server_test.go
@@ -1,4 +1,4 @@
-package rpc
+package server
 
 import (
 	"bytes"
