+++ pkg/rpc/server/testdata/test_contract.avm
